+++ src/com/test/payment/PaymentProcessor.java
@@ -37,7 +37,7 @@ public class PaymentProcessor {
             return;
         }
 
-        double amount = userBooking.getTotalPrice();
+        int amount = userBooking.getTotalPrice();
         System.out.println("결제하실 금액은 " + decimalFormat.format(amount) + "원 입니다.");
 
         // 잔액 확인 및 충전/취소 옵션 제공
@@ -89,9 +89,9 @@ public class PaymentProcessor {
         System.out.print("충전하실 금액을 입력하세요: ");
         String input = scan.nextLine();
 
-        double chargeAmount;
+        int chargeAmount;
         try {
-            chargeAmount = Double.parseDouble(input);
+            chargeAmount = Integer.parseInt(input);
             if (chargeAmount <= 0) {
                 System.out.println("충전 금액은 0원보다 커야 합니다.");
                 return;
